refac: Rename TimSeriesPoint -> EnergyTimeSeriesPoint (#1496)
