+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/Interfaces/CalculationResults/EnergyResultTests.cs
@@ -35,7 +35,7 @@ public class EnergyResultTests
         Instant anyPeriodStart,
         Instant anyPeriodEnd)
     {
-        var emptyTimeSeriesPoints = new EnergyTimeSeriesPoint[] { };
+        var emptyTimeSeriesPoints = Array.Empty<EnergyTimeSeriesPoint>();
         var exception = Assert.Throws<ArgumentException>(() =>
             new EnergyResult(anyId, anyBatchId, anyGridArea, anyTimeSeriesType, null, null, emptyTimeSeriesPoints, anyProcessType, anyPeriodStart, anyPeriodEnd, anyFromGridArea));
         exception.Message.Should().Contain("empty");
